website/requirements.txt
flask
librosa
tensorflow
Pillow
facenet-pytorch==2.5.2
torch
opencv-python
grad-cam
moviepy
openai_whisper
torchaudio
torchvision
transformers
ffmpeg
dlib
opencv_python
scikit-learn
firebase_admin
python-dotenv
pyrebase4